backend/requirements.txt
Django==6.0.3
django-allauth==65.13.0
python-dotenv==1.2.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
requests==2.32.5
PyJWT==2.12.1
cryptography==46.0.5
django-celery-beat==2.9.0

# ── Caching ──
django-redis==5.4.0

# ── Production dependencies ──
psycopg2-binary==2.9.12
whitenoise==6.12.0
dj-database-url==3.1.2
gunicorn==25.3.0
